Initial definition of glibc_core
CQID: WIND00343489

glibc_core is intended to be a minimal collection of packages
required to boot a system and log in at the console. It should be
as small as possible without relying on packages such as busybox,
dropbear, etc. Currently yocto tasks _do_ include busybox but this
definition is only a starting point.

If mc is removed from task-core-basic-utils, this generates an image of 29 MB.

Signed-off-by: Amy Fong <amy.fong@windriver.com>
Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>