image-debuginfo: add companion debuginfo filesystem generation
Using the wrlinux distro, the image-debuginfo is automatically included.  To
enabled the debuginfo companion filesystem to be generated, set
IMAGE_GEN_DEBUGFS="1".

The debuginfo companion filesystem is not a complete filesystem, it is made
up of only the changes required to add the dbg-pkgs PACKAGE_GROUP to the
regular filesystem image.

In order to use the companion files, you need to first extract the runtime
filesystem, and then overlayer the debug components on-top.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>