image-debuginfo: Fix install when packages already have -dbg components
CQID: WIND00367491

When generating a debug-fs, the -dbg components are processed and any
that are not already installed are added to the image.  During this process
a few commands are run, and the result codes need to be ignored in order
to avoid an error condition when there are no new components to install.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>