python: Fix python-nativesdk builds
CQID: WIND00365071

This is a temporary change to enable python-nativesdk builds with the WR
change that disabled the eglibc-nativesdk work.

The change works by disabling the BBCLASSEXTEND in the core python recipe,
and then provides a new python-nativesdk recipe based on the existing
python-native recipe.  This results in a nearly identical build, except
for the python-nativesdk-sqlite3 is not built.

Since that python-nativesdk-sqlite3 is not built, the external-python-tarball
needs to be modifed to no longer require it.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>