Fix path search for the wr-simics layer
CQID: WIND00364217

The WB runs the command

PROJ_DIR/scripts/config-target.pl list

Which will list the available simulators for a connection, and this
was not working properly because the simics layer search function was
not returning the correct value.

The simple solution turned out to be to use the same semantics as the
WR Linux 4.3 and and define the TOP_PRODUCT_DIR correctly in the
PROJ_DIR/Makefile and fix up config-target.pl to search for both
wrll-simics and wr-simics.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>