virtual/libgl: Make mesa-dri the preferred provider
CQID: WIND00369960

Make mesa-dri the preferred provider of virtual/libgl globally so there won't
be any issues of multiple recipes provided. In some cases mesa-xlib is the
preferred provider but in world builds xorg-xserver is built and it depends on
mesa-dri and they will both be built and end up providing virtual/libgl. This
is being done because on some targets the prefer mesa-xlib for virtual/libgl
and mesa-dri will be built for xorg-xserver and cause a multiple providers
error since both provide virtual/libgl

This was tested by doing world builds for a qemumips target. Before the changes
the multiple providers error was caused, after making mesa-dri the preferred
provider it wasn't caused anymore.