distro: default to not including a /etc/hostname file in the rootfs
CQID: WIND00361973

This is one of a series of commits made to several product trees to
allow for more flexible hostname behavior. By setting the value of
hostname to be 'none' we will not have a /etc/hostname file included
in the rootfs. This will result in a hostname passed on the command
line or via dhcp to be used and if neither of these are present the
default of 'localhost' will be used. This behavior can be overwritten
in the local.conf file.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>