Add ability to deal with non-absolute paths in layer_paths
CQID: WIND00415766

The configure script will now generate the default layer_paths file
without any absolute path information.  The config-target.pl and
create-usb.pl need to be aware of this and automatically prefix the
platform projects path while processing the file.

This allows the entire platform project to be relocated and all the
scripts will continue to work correctly.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>