config-target.pl: add screen console support
[WIND: TA17748]

GNU screen has a couple of minor peculiarities that means it cannot be
treated quite the same as xterm (et. al.) when being used as a qemu
console.  It turns out to be very nice to have this support, though, since
GNU screen tends to do the right thing when invoked on the command line.
That is, start a new session if none is running, start a new window in the
current session if one is.

Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>