create-usb.pl: Add advanced options and flexibility
It should be possible to specify a custom initrd for the usb boot
media.  Also if the end customer does not desire modifications
to anything in export/dist such as the ability to activate
read-only or read-write settings the export/dist directory
should be left alone by using "--pristine".

The "--initrd=<initrd>" will allow the specification of a custom
initrd or initramfs image.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>