wrlinux.conf: set default VIRTUAL-RUNTIME_syslog
CQID: WIND00441695

The usage to RPROVIDES virtual package is not correct since
there is no concept of PREFERRED_PROVIDER in the package manager
namespace, change to use VIRTUAL-RUNTIME_syslog instead.

we want to use rsyslog as the default syslog, but its license GPLv3,
so need a check if GPLv3 is allow, if not, use sysklogd.

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>