wrlinux-common.inc: remove IMAGE_GEN_DEBUGFS from BB_HASHBASE_WHITELIST
CQID: WIND00444986

The IMAGE_GEN_DEBUGFS is used for generating the debugfs, if we put it
in BB_HASHBASE_WHITELIST, then the change of IMAGE_GEN_DEBUGFS would
have no impact on the rootfs generation, and thus our:

make fs
make fs-debug

doesn't work correctly.

It worked in the past was becuase the rootfs always regenereted since it
had the nostamp, now the rootfs doesn't have nostamp any more, it only
regenerates when the hash changes.

Remove the IMAGE_GEN_DEBUGFS from BB_HASHBASE_WHITELIST will fix the
problem, only the rootfs depends on it, so it doesn't have any effect on
other recipes, which means that only the rootfs would regenerate when it
is changed, other recipes would not rebuild.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>