wrlinux-tiny.conf: Add PACKAGECONFIG to disable opkg-utils python dependencies
There was a failure to build wr-tiny, because perl was unexpected
added, the following is the runtime dependencies:
...
"packagegroup-wr-tiny" -> "update-alternatives-opkg"(opkg-utils)
"opkg-utils" -> "python"
"python" -> "openssl"
"openssl" -> "perl"
...

Since the wrlinux-tiny distro was based on poky-tiny, and poky-tiny
config has disabled python usage in opkg-utils. We should do the
same thing in wrlinux-tiny config.

Here is the related commit in poky:
...
commit b908caca1af1d4150eafd3d3001ea19599657270
Author: Richard Purdie <richard.purdie@linuxfoundation.org>
Date:   Fri Jan 17 11:40:29 2014 +0000

    poky-tiny.conf: Add PACKAGECONFIG to disable opkg-utils python dependencies

    (From meta-yocto rev: 5b191256efdf4ae7d0f58cd2612048d832b86556)
...

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>