wr-host-tools.inc: add gdb-cross
ISSUE: LIN7-1071

With the MG toolchain, cross gdb was a toolchain component and
just got copied in automatically. With the new toolchain, we have
to actually request gdb-cross if we want it. And there isn't
actually a "gdb-cross" package; you have to specify
"gdb-cross-${TARGET_ARCH}". Since this is slightly complicated for
a typical end-user, and we have lots of documentation which just
assumes that scripts/gdb.sh can find a working gdb, add it to our
default list of host tools.