ARM no longer needs -mapcs-frame
ISSUE: LIN7-1642 TCWRLGCC-5

Long ago, when we needed to rebuild things with -fno-omit-frame-pointer
for oprofile to work, we were advised to use -mapcs-frame in order to
have reasonably reliable stack backtraces. That was 2008. It is now 2015.
We no longer need -mapcs-frame, and indeed, it's considered an
obsolescent feature and can cause segmentation faults. Drop it.