Fix mcopy EFI files failed issue
Issue: SCP8-85

In last commit(fb8928e4), I have added the code to copy
files in export/dist/boot/efi/EFI/BOOT to vfat partition,
and the return code is check as usual. This will result
in the failure when the project is not using gurb-efi.
Because there is no this foler. In fact, we are just
care about coping them when they exist, so just ignore
the return code of copy command, assuming it wil always
right when there are these file.

Signed-off-by: Junxian.Xiao <Junxian.Xiao@windriver.com>