create-usb.pl: correct the formula of calculating parameters of a partition
Issue: LIN8-3484

Parameters of a partition is calculated too roughly,just by number of sectors
belonging to it. There should have been a little adjustment to them.

Of a partition,we can make an assumption,that index of start sector is SS,
that index of end sector is SE,that size of sector as SSIZE.

So,offset of end byte of this partition should be ((SE + 1) * SSIZE - 1),
not (SE * SSIZE).Offset of start byte is (SS * SSIZE),it's all right.
And count of all bytes in this partition should be (((SE + 1) * SSIZE - 1) -
(SS * SSIZE) + 1 ).

Signed-off-by: Qi Hou <qi.hou@windriver.com>