image-fs_finalize: add the explanation for WRL_FS_FINAL_PATH
Issue: LIN9-3373

If we set the WRL_FS_FINAL_PATH, the SRC_URI of the related iamge
will be set.
In fact, the recipes such as images and packagegroups don't actually
fetch or build any source, so there's really no point having
LIC_FILES_CHKSUM set, but when the WRL_FS_FINAL_PATH is used, the
LIC_FILES_CHKSUM of related image must be set.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>