conf/wrlinux-version.inc: Define new system wide version information
Issue: TA194347

We have defined the new version format to be:

      10.YY.WW.RCPL

Where 10 is the major product version
Where YY is the 2-digit year (such as 17)
Where WW is the work-week (such as 5, 42, 42a) [a = second release of week]
Where RCPL is the RCPL (0 means initial or CI version).

In addition, we also define the product 'branch'.  This is either
'development' (internal or continuous integration), 'base' the stable
free version or 'LTS' the commercial version.

The version information is appended to the top of the standard build info,
so it appears at the top of a build log such as:

Build Configuration:
WRLINUX_VERSION   = "10.17.37.0"
WRLINUX_BRANCH    = "development"
BB_VERSION        = "1.35.0"
BUILD_SYS         = "x86_64-linux"
NATIVELSBSTRING   = "universal-4.8"
DISTRO            = "wrlinux"
DISTRO_VERSION    = "10.17.37.0"
...

This is intended to make it easier for use to understand what the user
is doing, even if they are not using a 'wrlinux' derived DISTRO value.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>