wrlinux-version.inc: Adjust the default branch value based on UPDATE
An UPDATE value of '0', indicates this is a development branch.  A non-zero
value would be that it is either Base or LTS.  We want to define Base as
the default value in this case.  The wrlinux-x (setup) will adjust the
generated local.conf to provide 'LTS' when appropriate.

Also a minor whitespace cleanup to the UPDATE field.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>