wrlinux-common.inc: disable the dnsmasq service
Issue: LIN10-1402

The service will fail to start/stop when dnsmasq and
bind (provides named) are both installed since they
are using the same port 53:

| dnsmasq: failed to create listening socket for port 53: Address already in use

So disable the service by default, if the user really
want to use two dns servives, the config needs to be
changed to use different interfaces or ports.

As reference, other distro like centos/fedora also
disables the services by default on installation.

Upstream notes:
The patch was sent upstream meta-networking but rejected,
It is suggested to add in the distro level.

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>