kernel-dev: carry forward the kernel-dev layer from WRL9
Issue: LIN10-2101

Derived from WRL9 (wr-kernel, HEAD commit 7487a6a5b4cb) and should be
mostly functionally equivalent.

Without wr-kernel and the linux-windriver there is no concept of
kernel-headers so the parts that used to exist to create
KERNEL_ALT_HEADER_DIR have been dropped. This is also consistent with
recipe specific sysroots (RSS) as populating the sysroot with these
headers no longer accomplishes the goals of WIND00392765 (user visible
headers).

Since we are now using linux-yocto we dropped linux-libc-headers as we
can simply rely on Yocto standard infrastructure to provide this.

The README has been updated to include descriptions of the three
template features. The linux-yocto-custom.bbappend can be used as an
example of using an user supplied kernel repo and defconfig. All 3
templates are functional out of the box, and have been tested.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>