linux-yocto: Ensure the proper URI and latest branch commit are used
Issue: LIN10-2308

We use an anonymous python command to adjust the SRC_URI to ensure that the
repository provided by this layer is used instead of the Yocto Project layer.

We also adjust the QEMU BSP settings to ensure they track the top commit in
whatever branch they are defined to used, instead of a specific commit.

Both of these changes make it easier to maintain the QEMU BSPs when we uprev
the kernel branches in the Wind River copy of the yocto kernel trees.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>