linux-yocto-wrlinux: avoid project path in computing checksum
Issue: LIN10-2424

We have project path in SRC_URI in linux-yocto, and this has negative
effects including not being able to share sstate cache and having
signature mismatch error when using eSDK to build kernel module.

To fix this, instead of using the actual path to replace SRC_URI,
we use "${LAYER_PATH_wrlinux}". In addition, we add this variable
to BB_HASHBASE_WHITELIST so that its value, which contains project
path, would not be evaluated when computing checksum.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>