Move the kernel-dev template and associated class to 'wrlinux'
Issue: LIN10-2595

We want the ability for the user to include the kernel-dev and kernel-devsrc
packages into their generated SDK.  This template is moved from the unsupported
wrlinux-kernel-dev layer to the supported 'wrlinux' layer and is available
for all configurations.

Also update the wrlinux-kernel-dev README file, and create a basic README
for the template.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>