template/feature/kernel-dev: Update to new method
Issue: LIN10-2595

The KERNEL_ALT_HEADER_DIR variable is not yet implemented in this version.

The variable also has a different purpose then inclusion of the kernel sources
into the SDK.  Remove it, it may be added in the future when this functionality
is restored to the wrlinux-kenrel-dev layer.  (It would be added in a new
template.)

Add a workaround to include perl, python-dev, bash and gawk due to a problem
with DNF requirement exclusions.

Remove the kernel-sdk.bbclass:  The user can easily install the scripts based
on a standard set of documented steps.  This is much easier and safer then
automating the process

README: Add steps explaining the usage of the kernel in an SDK

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>