linux-yocto-*: Add appends for linux-yocto-rt and linux-yocto-tiny
With the new linux-yocto based system, a kernel can be selected using
PREFERRED_PROVIDER_virtual/kernel = "linux-yocto-rt"

If this is done, a user may still want the components from this layer. So
we move the previous bbappend into a unique .inc file and then add overrides
for the standard linux-yocto variants.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>