recipes-kernel/linux/linux-yocto-wrlinux.inc: Add KTYPE validation
Issue: LIN10-2601

In many cases we want to verify that the selected LINUX_KERNEL_TYPE is
supported via the TARGET_SUPPORTED_KTYPE variable that may be defined by
a BSP.

If TARGET_SUPPORED_KTYPE -is- define, we verify that the LINUX_KERNEL_TYPE
(only when using the standard linux-yocto packages) is in the list of
supported ktypes.  If it is not, we set the package to 'skip' with a
corresponding message telling the user that KTYPE is not supported.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>