gen-linux-srcrev.sh: Update to define the LINUX_VERSION per KBRANCH
Issue: LIN10-2726

Since the KBRANCH can be on different versions, move to defining the
LINUX_VERSION on a per KBRANCH basis.  We still use the meta data 'KVER' to
defined the default LINUX_VERSION.

Also, no longer carry the 'rebase' branches in the srcrev files.  They are
intended to be temporary and should not be used by any BSP.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>