wrlinux-ovp: lmplement host and guest kernel based on [ LIN10-2121 ] linux-yocto-rt
Issue: LIN10-2121

The default kernel will be the host kernel.  It relies on an x86-64
BSP that supports preempt-rt.

A message will be generated informing the user that they are building the
default host kernel, and they need to enable a template to build the
guest kernel.

The template, feature/kernel-ovp-guest, should be used to enable the guest
ovp kernel.  This kernel has the restriction of being designed for use with
'qemux86' or qemux86-64' machines only.  This is enforced by
COMPATIBLE_MACHINE.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>