lmbench: fix lmbench script syntax error
Issue: LIN1018-148

When running lmbench test script, we meet the following error:

    /opt/benchmark/os/wr-lmbench/scripts/utility.sh: line 42: break: only
meaningful in a `for', `while', or `until' loop

This is because in wr-lmbench-test.sh script, calling a while loop, and the
while loop calling a function, this function was defined in utility.sh, when
the function satisfied some cases, calling the "break" to end the while loop.

The solution was that the while loop use the return code of that function,
when the code meet someone, the while loop calling the "break".

Signed-off-by: Dengke Du <dengke.du@windriver.com>