wrlinux-ovp: enable the VFIO feature for OVP10
Issue: LIN10-3485

When creating VM with pci-assign, it failed with the following error.

'''
qemu-system-x86_64: -device pci-assign,host=0000:07:10.0,id=mydevice: No IOMMU
found. Unable to assign device "mydevice"
'''

Legacy PCI device assignment has been removed in linux-4.12 and the community
suggested using VFIO as a substitute. So we should use VFIO instead of
pci-assignment.

But the VFIO driver was not compiled in OVP10 kernel. So add the supprort of VFIO.

(LOCAL REV: NOT UPSTREAM) --Inappropriate [enable feature]

Signed-off-by: Liu Haitao <haitao.liu@windriver.com>