buildtools-tarball: remove nativesdk-perl package
Issue: LIN10-4357

The eSDK building process stucks at sanity-check for missing perl modules.
The root cause is that we believe that buildtools-tarball is inherently a
tool set instead of a fully functional SDK, so the variable OECORE_NATIVE_SYSROOT
is not set, which causes @INC contains wrong perl module path, and then
perl can not find module. Because we do not need perl in buildtools since
WRL10.17, it would be fine to replace it with nativesdk-buildtools-perl-dummy
(provides perl, but doesn't really install anything).

Signed-off-by: Rui Wang <rui.wang@windriver.com>