libvirt: Modify existing variables instead of appending
Issue: LIN10-4771

Older versions of libvirt allowed us to append configuration variables
to the end of files like libvirtd.conf, such that if there were
multiple times a value was set the last one found in the file would be
used. Now either the reverse is true or the search stops after the
first occurance of a variable assignment is used. As such we need to
modify the existing assignments in the conf files, instead of simply
tacking on to the end of the configuration file. Without this change
the default 'libvirt' group is attempted to be used and this will
cause libvirtd to fail to start:

  error : virGetGroupID:1125 : invalid argument: Failed to parse group
  'libvirt'

With this change we go back to using 'qemu'.

(LOCAL REV: NOT UPSTREAM) --backport from WRL1018
[commit:8fd83683eccfda30d1cab8c1c45d707e7db1f018]

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>