dpdk: fix compiling failure on qemux86-64 and qemux86
It probably wouldn't even work since default value is core2duo
for qemu in tune-core2.inc which doesn't support SSE 4.2 (which is a
requirement for dpdk).

So use native value for qemux86-64 and qemux86 although these two
arches does not support SSE 4.2, but it still makes sense to do the
build testing on Yocto supported machine.

Upstream meta-dpdk maintainer suggests `this be done in your local
configuration'

Issue: LIN1018-2606

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>