nxp-imx8: change the priority of EXTRA_OECONF in wrlinux layer
RallyID: US113230

The variable of EXTRA_OECONF in wrlinux layer should be the lowest priority
because it is only used for the scene that has no hardware acceleration. In
this scene, framebuffer should be the backend for weston. But once using
GPU in imx8mq, drm should be the backend. So we should let the EXTRA_OECONF
in weston_3.0.0.bb in imx8-graphic layer to overrides the variables here.
So change the assignment from "=" to "??=".

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>