wrlinux-image-initramfs: add the libgcc into the ramfs
Issue: LIN10-5363

Add the libgcc package into the initramfs, since the multi-thread
need it. But we use PACKAGE_INSTALL instead of IMAGE_INSTALL to limit
the ramfs size in the initramfs bb file. So that the PACKAGE_INSTALL
value will override the install_packages. There is no the libgcc package.

If the customer want to add the other package, they can use the variable
IAMGE_INSTALL_INITRAMFS_append in the local.conf.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>