container-images: make a version of c3-*-containers available in wrlinux
Issue: LIN1018-3473

There is currently no baseline container image definitions available
in oe-core or base wrlinux so make a version of the c3-*-container
image recipes and helpers available in wrlinux.

In upcoming releases of Yocto we anticipate that a shared mechanism
will be available to facilitate the generation of container images
oe-core but for LTS18 this work was only in the planning state in the
Yocto community. We do, however, have helper files for container image
creation in meta-overc which we used here to derive the additions
included in this commit to make this functionality available outside
of overc.

By making a form of these files available outside of overc we can make
use of the wrlinux or other distros which have broader whitelists.

Although container images are not all that different than regular
images produced by bitbake, by making these helpers available we
provide easier ways to define the contents of the images and also
provide a central location for making optimizations and 'tweaks' to
the images which all images that derive from these helpers will
benefit from.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>