templates: Add additional in-development templates
As the system is developed, new versions of certain components are introduced
that may be incompatible or too risky to make the default versions.  Add
a couple of additional templates to enable these in development components.

linux-yocto-dev - Add README to explain what it's for
toolchain-next  - Template to enable in-development toolchain
userspace-next  - Template to enable specific in-development userspace items

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>