wrlinux-distro: Resync distro components to oe-core
This resyncs the distro settings to oe-core.  Both before and after this
change result in identical DISTRO_FEATURES definitions.

The new way of doing things defines a WRLINUX_DEFAULT_DISTRO_FEATURES which
defines the basic features, not otherwise controlled by the BACKFILL system.

Note: 'ldconfig' is in this list, because currently the system includes
ldconfig in all configurations -- even though this is a bug.  It will be
corrected in subsequent commits.

Additionally the DISTRO_EXTRA_RDEPENDS for qemu, has been changed to use
'qemuall' override.  This will ensure no matter what qemu BSP is used the
settings will be applied without specific knowledge of each possible qemu
BSP.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>