Wayland: add weston framebuffer backend support for nxp-imx6
RallyID: US129464

The SDK doesn't provides GPU acceleration librararies (vendor's libEGL.so)
for wayland for imx6. And weston can't use "drm" backend with the generic
"libEGL.so" which is provided by mesa because it can't provide some attributes
which make weston work well. So we use "framebuffer" instead of "drm".

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>