dpdk: Make the PREFERRED_VERSION_dpdk more specific
Issue: LINUXPG-21

We plan to add a 18.11.1 bb for the dpdk package. This version is
supposed to be used by some specific BSPs only. And we would like
to leave the default version of dpdk as 18.05 in order to avoid
introduce some breakage on the current machines. So make the
PREFERRED_VERSION_dpdk more specific to 18.05.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>