wrlinux-distro-settings.conf: Add workaround for uid host-user-contaminated issues of glibc-locale
Issue: LIN1018-4223

There are warnings similar with below reported occasionally:
WARNING: QA Issue: glibc-locale:
/glibc-binary-localedata-dv-mv/usr/lib/locale/dv_MV/LC_NAME
is owned by uid 20853, which is the same as the user running bitbake.
This may be due to host contamination [host-user-contaminated]

It is because of rare race condition in the processing of the user ids.
This is a known issue for yocto and it is hard to reproduce.

We have seen it in other recipes besides glibc-locale, even though
it is incredibly rare. Because glibc-locale is the main module that this
issue is reported on, just turn its warning into error to ask the user to
pay attionention to the workaround way in the printed information.

Signed-off-by: Li Zhou <li.zhou@windriver.com>