saftest: fix case failure for saHpiDomainTagSet/4-2
Issue: LIN1019-1656

In case saHpiDomainTagSet/4-2, it considers 0xFFFF to be an invalid
character and sets it to the first 2 bytes in the data of DomainTag.
Unfortunately, this is not correct because 0xFFFF is an valid character
in UTF-16 Basic Multilingual Plane (BMP). It should set an invalid
surrogate pair between 0xD800-0xDFFF. In this case, we set it to 0xDBDB,
this is an illegal surrogate pair and the case can return the expected
result.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>