vrf: fix zebra startup failure
Issue: LIN1019-1657

We don't need to create /var/run and /var/lock in rootfs layout since
they should be symbolic links of /run and /run/lock. The
systemd-tmpfiles will create them when booting.
Since systemd 240, it will check pid file of service in /run even if the
PIDFile of service points to /var/run. So the separated directories /run
and /var/run will make systemd can not check the pid file and get the
correct status of services.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>