refpolicy: fix file context of /usr/libexec/udevadm
Issue: LIN1019-1731

The previous fix of LIN1019-1616 is not correct. The hard link files
that have different contexts are labeled based on the last matching
specification. If we remove the file context of /usr/libexec/udevadm and
it is in last match. It will be labeled as bin_t which inherit from the
context of /usr/libexec directory. So we should keep it with the same
file context of /bin/udevadm.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>