sysstat: fix PACKAGECONFIG and move to bbappend file
Issue: LIN1019-1712

Add cron to PACKAGECONFIG. Filter systemd distro feature in PACKAGECONFIG.
In this way, 'sar' could work as expected.

The PACKAGECONFIG setting for sysstat was in a bbappend file,
holding the value of 'lm-sensors'. It was moved to conf file because
at that time lmsensors was not upstreamed.

Now the lmsensors recipe has been upstreamed to meta-openembedded.
We should move it back to bbappend file.

Also, as 'lm-sensors' setting is WRLinux specific and cannot
be merged in oe-core, we need this bbappend file.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>