libvirt: Move to tls mode for default remote connection
Issue: LINUXEXEC-1942

Asked by customer, we change the default link mode to a secure mode,
tls(SSL 3.1) which is the same as libvirt upstream.

Since v1.3.5 used by WRL9 to current v5.4.0, tcp is disabled and tls is enabled
by default. But meta-virtualization overwrites the configuration to enable tcp
and disable tls, without any explanation in
8378d87ced87 ("libvirt: Initial package creation.").

Signed-off-by: He Zhe <zhe.he@windriver.com>