umip: add recipe
Issue: LINUXEXEC-2952
Issue: LIN1019-1655

The umip has been removed in mete-openembedded with commit
bb4aa80bfbd67ddfafeaf7700aff993592387a21 because it depends on
ipsec-tools which has been dropped. But in fact the ipsec-tools is not
the build dependency or essential runtime dependency for it.
It is also supported by other ipsec tools. e.g strongswan
(https://wiki.strongswan.org/projects/strongswan/wiki/MobileIPv6).

The upstream rejected to add it back because the package is unmaintained
since 2013. But it is the only implementation of Mobile IPv6 and NEMO
Basic Support for Linux and is required by CGL. So we add the recipe
back.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>