ovp: Disable docker to be compiled to lib32-docker
Issue: LIN1019-2290

When building wrlinux-image-ovp-kvm with general template feature/docker, we
would get the following failure.

ERROR: Nothing RPROVIDES 'lib32-docker' (but virtual:multilib:lib32:
/build/layers/wrlinux/wrlinux-distro/recipes-base/images/wrlinux-image-std.bb
RDEPENDS on or otherwise requires it)

This is because OVP host image only supports x86-64, but the template does
not impose the limit and would trigger the multilib extension, and then
lib32-docker fails to be generated.

This patch put docker into NON_MULTILIB_RECIPES.

Signed-off-by: He Zhe <zhe.he@windriver.com>