avl.28.3: refactor the testcase
Issue: LIN1019-2481

when fsck system, recover journal only happened when the file
system is need recovery, but previous way cannot ensure this.

For the old way,  first dd something on the filesystem to make
it busy, and then umount with -fl, so there will be a short time
that filesystem is detached from the filesystem hierarchy, but
filesystem are not cleanup. It hopes to fsck system in this
short time, but this short time is not controllable.

refactor to make sure that when fsck system, the filesystem
needs recovery

Signed-off-by: Changqing Li <changqing.li@windriver.com>