wrlinux-overc: add to 'wr-init' to whitelist to fixup build error
Issue: LIN1019-3044

We started seeing the following errors:

    WARNING: wr-init is not whitelisted, figuring out PNWHITELIST...
    ERROR: Nothing RPROVIDES 'wr-init' (but
    .../layers/oe-core/meta/recipes-core/packagegroups/packagegroup-core-boot.bb
    RDEPENDS on or otherwise requires it) wr-init was skipped: Not
    supported in this configuration by Wind River. To override, add to
    your local.conf: PNWHITELIST_wrlinux-distro += 'wr-init'

This is caused by commit b0ae6d9bbef6 [wrlinux-overc.conf: add
osv-wrlinux to distro overrides] as it had the side effect of "turning
on" 'RDEPENDS_${PN}_append_osv-wrlinux = " wr-init"' in
packagegroup-core-boot_wrlinux.inc which in turn created an
unbuildable dependency chain since 'wr-init' is not whitelisted for
the 'wrlinux-overc' distro.

The 'wrlinux' distro already has these packages in its whitelists but
since the 'wrlinux-overc' distro maintains its own whitelist we need
to add 'wr-init' to get things back to a buildable state. NOTE that
for OverC the 'wr-init' is a nop since it only has an effect systems
with sysvinit.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>