wrlinux-overc: default to use genericx86-64
Issue: LIN1019-2969

Our default build we use in the official documentation and other
places for OverC uses the genericx86-64. We should therefor set the
default INSTALL_BSP to "genericx86-64" so that our docs work
out-of-the box. Other BSPs will cover this in their specific READMEs
in the docs directory, as we have done with README.qemuarm64.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>