wrlinux-overc: kubernetes: work around multilib/whitelist issue
Issue: LINUXEXEC-4903

When attempting to build cube-k8s-node we are getting an error:
    ERROR: Required build target 'cube-k8s-node' has no buildable providers.
    Missing or unbuildable dependency chain was: ['cube-k8s-node', 'lib32-cube-k8s-node', 'lib32-packagegroup-k8s-compute', 'lib32-kube-proxy']

This is caused by the whitelisting system in place. When building
cube-k8s-node with the Yocto Project, or with Wind River Linux, we
don't actually ever attempt to build lib32-cube-k8s-node. It actually
isn't possible in either YP or WRL to build. However, the vetting of
packages against the whitelist attempts to parse the multilib variants
of the kubernetes packages which causes the above ERROR. We can work
around this by excluding kubernetes from the multilibs.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>