wrlinux-overc: wrlinux-overc.conf: add recipes that can skip lockdown
Issue: LINUXEXEC-4903

Following the documentation to run overc with QEMU we need to allow
for the building of ovmf for 'bios.bin' as we only build an EFI
enabled image. We can't generate sigs for these recipes as they are
not available for ARM and our sig gen tooling can't handle this
currently, so our only option is to skip them, which is fine since
they don't get included in any image.

Add 'ovmf' and its dependencies to SIGGEN_UNLOCKED_RECIPES.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>