feature/ostree: Move feature/ostree to the top level
We want to be able to use feature/ostree with any rootfs and distro.
The code is generic and this will be how we add the meta-overc
support.

This commit allows the distro wrlinux-overc to build with
feature/ostree, but only when the overc hashes are disabled to enable
dev mode.  A new series of patches will complete the functionalit
later, but this will create the ability to develop the feature.

[ Issue: LINUXEXEC-5186 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>