feature/encrypted-storage: Fix feature/ostree dependency
Ostree knows how to deal with encrypted volumes and must manage the
file system layout directly.  When using feature/ostree that WKS file
must be used.

The work around was to make the feature/ostree appear last in the
template list, this patch fixes it such that the template inclusion
order will no longer matter.

[ Issue: LIN1019-3802 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>