qemu: remove patch about NMI
Issue: LIN1019-4000

remove qemu-inject-NMI-in-i6300esb-watchdog-timer.patch
the patch of function is:
when i6300esb watchdog timeout, it will inject NMI to guest OS.
in the version of qemu, the function was implemented in the codes.
furthermore, the patch cause qemu crash, too.

add the function in the version qemu:
qemu option:
-watchdog-action inject-nmi
libvirt xml config:
<watchdog model='i6300esb' action='inject-nmi'>
</watchdog>

Signed-off-by: Li Wang <li.wang@windriver.com>