refpolicy: fix ssh login fail to chdir to /root
Issue: LINCD-1187

when ostree is enabled, root dir becomes a symlink file, it would
not get the right label for such a kind of root, then sshd does not
have permission to access symlink file which will led can not
chdir to /root, so label symlink root as the correct label to
fix this.
The alias in .subs_dist prevent label the symlink root, so define
it in usermange.fc.

Signed-off-by: Liwei Song <liwei.song@windriver.com>