distro/wrlinux.conf: Set the default kernel for all the wrlinux distros
Issue: LINCD-1274

In the current code, the setting of the PREFERRED_VERSION_linux-yocto
sprinkle over the BSP's machine conf, but it seems pretty unnecessary
to foist this same info into every BSP. So we should move this into a
common place, such as the distro conf. This is also what the Poky
distro does. Another bonus is that we can easily get what kernel
version is used by the CD standard. At the same time, drop the
stale PREFERRED_VERSION_linux-yocto setting in wrlinux-tiny.conf.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>