kernel-initramfs-image: solve missing bundled image issue
Issue: LINCD-1614

When setup with
"--templates feature/initramfs-integrated feature/initramfs-install",
the bundled kernel and initramfs file isn't installed to rootfs because
do_install of kernel-initramfs-image is running after do_deploy of kernel.
Add the right dependency for this.

Signed-off-by: Li Zhou <li.zhou@windriver.com>

It's not necessary to depend on virtual/kernel
There is no do_unpack for kernel-initramfs-image
It's not enough and not necessary to require virtual/kernel:do_deploy to
finish before do_populate_lic

Signed-off-by: Zhang Qiang <qiang.zhang@windriver.com>