CUT: use ip command and update ping6
Issue: LINCD-2470

1,
After the below commit applied in oe-core, the net-tools which
provides ifconfig changes a little.
dacfd69506 net-tools: backport a patch from upstream to use the same ifconfig format as debian/ubuntu

Switch to use ip command to get the interface address to fix the
gap as the ifconfig command is more and more deprecated.

2, use "ping -6" to instead of the ping6.

3, remove the redundant logic which used to get the avaiable
ipv6 enabled interface as we already have set the ipv6 enabled
interface LOCAL_ETH.

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>