third-party-dl.bbclass: Add it
Issue: LINUXEXEC-9524

Include notification during build time, when some 3rd party components
might be added or pulled as dependency.

Add a prefunc for do_fetch. When the source is downloaded, check if it
is supported by wrlinux. If not, print a warning message.

The supported recipes list is from wrlinux-recipes-list.inc which
resides in wrlinux-distro/conf/ and please update it when there is
new recipes or remove some ones.

Signed-off-by: Changqing Li <changqing.li@windriver.com>