whitelist.bbclass: don't check the depends' templates
Issue: LINUXEXEC-1559

For the recipes's template check, take ade for exmaple.
Previously, we not only check the templates which ade belongs,
but also check the templates for the recipes which ade depends.
$ bitbake -cfetch ade

Actually, it's superfluous and not correct. We only need to check
the templates which ade belongs and the templates which whitelist
the ade should already whitelisted the recipes which ade depends.

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>