wrlinux-kernel-dev: enable SDK related patches only for wrlinux-kernel-dev
Issue: LINCD-4377

WRLinux only support bulding kernel modules inside SDK when wrlinux-kernel-dev
is enabled. So it's reasonable to move these two patches here. In this way, if
kernel is updated, and the patches don't apply, we don't block the whole testing.

In addition, we only enable these patches for linux-yocto for now. If linux-yocto-dev
also needs it, we may need to make extra two patches for it, as the Makefile may diff
between the standard kernel and the next kernel.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>