wrlinux-image-minimal.bb: fix image features removal
Issue: LINCD-5005

It appends 'x11-base' to IMAGE_FEATURES for non-container images in an
anonymous function in wrlinux-bin-image.inc which is applied after
evaluation of EXTRA_IMAGE_FEATURES. Then the '_remove' does not work as
expected and 'x11-base' is not removed from IMAGE_FEATURES eventually.
So apply '_remove' to IMAGE_FEATURES directly to make sure the image
feature is removed indeed.

wrlinux-image-full is not affected that it includes 'x11-base' for
non-container images.

Signed-off-by: Kai Kang <kai.kang@windriver.com>