feature/dhcp-server-kea: add template for kea
Issue: LINUXEXEC-13919

Currently the kea is install by default in small/std/cgl images.
However, its build time is a bit long since it is heavy. Considering
that it is not used frequently, we can remove it from the default
installation list to save build time and add a template for it. Then
users can install it by enable this template.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>