scap-security-guide: accounts_remove_unneeded: fix games account detection
Issue: LIN1019-6800

The test should fail if "games" account is present on the system, but currently
it does the opposite: it passes if "games" account is present in /etc/passwd.
Adjust the oval logic to match the intended behavior and also provide a
remediation script to delete the user if necessary.

Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>